secret_santa/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-anymail==8.6
django-environ==0.9.0
django-extensions==3.2.1
django-fieldsignals==0.7.0
django-model-utils==4.3.1
django-render-block==0.9.2
django-templated-email==3.0.1
idna==3.4
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
